requirements.txt
fastapi
uvicorn
canvasapi
python-dotenv
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
